Huge overhaul of how comms with page are handled.
To overcome Manifest V3 service workers' 5-minute activity limit (see
https://bugs.chromium.org/p/chromium/issues/detail?id=1152255) we
instead inject an iframe into the host page (beta.timing71.org) and use
cross-window messaging to communicate. WebSocket connections are started
in that iframe, which uses the host.html entry point and has the
cross-origin permissions of an extension background script.

We still need background.js for a bunch of things, so it stays around.