Move all database access to extension-side code.
We can't access the IndexedDB data from host.js if third-party cookies
are disabled in Chrome - for some reason this is included in the
restrictions for that setting, so it can't be accessed from `host.js`.
We have to do a little song and dance to then get generated data (such
as the replay ZIP files) back over to host-side so they can be
downloaded, because extension-side code can't create object URLs to feed
into `FileSaver`.

Related: https://bugs.chromium.org/p/chromium/issues/detail?id=1224027